[typespec-ci] Improve change detection (#24825)
- Fixes #24813
- Fixes #24818

+++ eng/scripts/Validate-TypeSpec.ps1
@@ -8,18 +8,26 @@ param (
   [string]$SourceBranch
 )
 
-$typespecFolders = @()
+$exitCode = 0
 
 $typespecFolders = &"$PSScriptRoot/Get-TypeSpec-Folders.ps1" "$SpecsRepoRootDirectory" "$TargetBranch" "$SourceBranch"
 
-$exitCode = 0
+Write-Host "typespecFolders:"
 foreach ($typespecFolder in $typespecFolders) {
-  npx --no tsv $typespecFolder 2>&1 | Write-Host
-  if ($LASTEXITCODE) {
-    $exitCode = 1
+  Write-Host "  $typespecFolder"
+}
+Write-Host
+
+if ($typespecFolders) {
+  $typespecFolders = $typespecFolders.Split('',[System.StringSplitOptions]::RemoveEmptyEntries)
+  foreach ($typespecFolder in $typespecFolders) {
+    npx --no tsv $typespecFolder 2>&1 | Write-Host
+    if ($LASTEXITCODE) {
+      $exitCode = 1
+    }
+    git restore .
+    git clean -df
   }
-  git restore .
-  git clean -df
 }
 
 exit $exitCode
